# Speedup of the ebrisk calculator