--- a/debian/changelog
+++ b/debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Speedup of the ebrisk calculator
   * Extended the `minimum_intensity` feature to the classical calculator
   * Solved a memory bug when using the nrcan site term: due to a deepcopy
     the engine could run out of memory in the workers for large site collections