Continuation of https://github.com/gem/oq-engine/issues/5454. Now I managed to get a visible speedup for the Italy model on my workstation (more than 7x in aggregating losses).
Before
```
calc_48623                                  time_sec memory_mb counts 
=========================================== ======== ========= =======
total ebrisk                                23_329   64        4_260  
aggregating losses                          8_173    0.0       324_541
computing risk                              6_827    0.0       324_541
getting assets                              3_539    110       3_851  
EbriskCalculator.run                        2_428    2_146     1      
getting crmodel                             1_263    1.45312   3_851  
getting hazard                              228      0.0       11_496 
```
After
```calc_49065                                  time_sec memory_mb counts 
=========================================== ======== ========= =======
total ebrisk                                16_287   63        4_260  
computing risk                              6_990    0.0       324_541
getting assets                              3_561    110       3_851  
EbriskCalculator.run                        1_718    2_193     1      
getting crmodel                             1_304    1.44922   3_851  
aggregating losses                          1_163    0.0       324_541
getting hazard                              237      0.0       11_496 
```
Note for @raoanirudh : one can get over an order of magnitude speedup simply by reducing the number of concurrent tasks from 1600 (the value that was set in the global risk model file) to 20:
```
calc_49081                                  time_sec memory_mb counts
=========================================== ======== ========= ======
total ebrisk                                1_528    72        51    
computing risk                              1_060    0.0       48_234
EbriskCalculator.run                        250      1_899     1     
getting hazard                              205      0.0       11_496
aggregating losses                          131      0.0       48_234 
getting assets                              40       110       48    
```